+++ app/Http/Controllers/Tweets/TweetController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTweetFormRequest;
 use App\Http\Resources\TweetResource;
 use App\Models\Tweet;
+use Illuminate\Http\Request;
 
 class TweetController extends Controller
 {
@@ -22,14 +23,16 @@ class TweetController extends Controller
 
         $tweet->attachMediaFiles();
 
-        return $this->responseWithMessage("successful tweet");
+        $tweet->load(["user.profileImage", "media"])->loadCount(["replies", "retweets", "likes"]);
+
+        return $this->responseWithResource(TweetResource::make($tweet));
     }
 
     public function show($tweetUuid)
     {
         $tweet = Tweet::where('uuid', $tweetUuid)
                 ->with(["user.profileImage", "media"])
-                ->withCount(["retweets", "likes"])
+                ->withCount(["replies", "retweets", "likes"])
                 ->first();
 
         if (!$tweet) {
@@ -39,15 +42,22 @@ class TweetController extends Controller
         return $this->responseWithResource(TweetResource::make($tweet));
     }
 
-    public function destroy($uuid)
+    public function destroy(Request $request, $uuid)
     {
+        $user = $request->user();
         $tweet = Tweet::where("uuid", $uuid)->first();
+
 
         if (!$tweet) {
             return $this->responseWithMessage("the tweet does not exist or has already been deleted", 404);
         }
 
+        if ($tweet->user_id !== $user->id) {
+            return $this->responseWithMessage("you do not have permission to perform this action", 403);
+        }
         $tweet->delete();
+
+        $tweet->detachMediaFiles();
 
         return $this->responseWithMessage("tweet removed");
     }
